Can run in a console now
The program will run in a console now but you need to change the input
file paths manually
-> find the input file paths on your computer and paste the path into
bulkclass
-> make sure you do this both to AddMembers and FillItemArray

Next update will be making the paths to the input files self finding

+++ Project1/OtherFunctions.cpp
@@ -15,7 +15,7 @@
  * 	This function will return the pick to main. if not valid, will return
  * 	a -1.
  *************************************************************************/
-float GetAndCheckInt(int int1, //IN -  lowest allowable parameter for
+int GetAndCheckInt(int int1, //IN -  lowest allowable parameter for
 		//      correct values that need to be
 		//      error checked.
 		int int2) //IN -  highest allowable parameter for
@@ -25,7 +25,7 @@ float GetAndCheckInt(int int1, //IN -  lowest allowable parameter for
 
 	ostringstream buffer; // CALC - used for spacing purposes
 	bool inputOk;         // CALC - error checking bool
-	float pick;           // OUT  - will output a correct number
+	int pick;           // OUT  - will output a correct number
 
 	//OUT - error checking command
 	pick = 0;
